+++ config/openshift/v4_22/translate.go
@@ -22,7 +22,7 @@ import (
 	cutil "github.com/coreos/butane/config/util"
 	"github.com/coreos/butane/translate"
 
-	"github.com/coreos/ignition/v2/config/v3_6/types"
+	"github.com/coreos/ignition/v2/config/v3_5/types"
 	"github.com/coreos/vcontext/path"
 	"github.com/coreos/vcontext/report"
 )
@@ -104,7 +104,7 @@ func (c Config) FieldFilters() *cutil.FieldFilters {
 // can be tracked back to their source in the source config.  No config
 // validation is performed on input or output.
 func (c Config) ToMachineConfig4_22Unvalidated(options common.TranslateOptions) (result.MachineConfig, translate.TranslationSet, report.Report) {
-	cfg, ts, r := c.Config.ToIgn3_6Unvalidated(options)
+	cfg, ts, r := c.Config.ToIgn3_5Unvalidated(options)
 	if r.IsFatal() {
 		return result.MachineConfig{}, ts, r
 	}
@@ -161,11 +161,11 @@ func (c Config) ToMachineConfig4_22(options common.TranslateOptions) (result.Mac
 	return cfg.(result.MachineConfig), r, err
 }
 
-// ToIgn3_6Unvalidated translates the config to an Ignition config.  It also
+// ToIgn3_5Unvalidated translates the config to an Ignition config.  It also
 // returns the set of translations it did so paths in the resultant config
 // can be tracked back to their source in the source config.  No config
 // validation is performed on input or output.
-func (c Config) ToIgn3_6Unvalidated(options common.TranslateOptions) (types.Config, translate.TranslationSet, report.Report) {
+func (c Config) ToIgn3_5Unvalidated(options common.TranslateOptions) (types.Config, translate.TranslationSet, report.Report) {
 	mc, ts, r := c.ToMachineConfig4_22Unvalidated(options)
 	cfg := mc.Spec.Config
 
@@ -181,21 +181,21 @@ func (c Config) ToIgn3_6Unvalidated(options common.TranslateOptions) (types.Conf
 	return cfg, ts, r
 }
 
-// ToIgn3_6 translates the config to an Ignition config.  It returns a
+// ToIgn3_5 translates the config to an Ignition config.  It returns a
 // report of any errors or warnings in the source and resultant config.  If
 // the report has fatal errors or it encounters other problems translating,
 // an error is returned.
-func (c Config) ToIgn3_6(options common.TranslateOptions) (types.Config, report.Report, error) {
-	cfg, r, err := cutil.Translate(c, "ToIgn3_6Unvalidated", options)
+func (c Config) ToIgn3_5(options common.TranslateOptions) (types.Config, report.Report, error) {
+	cfg, r, err := cutil.Translate(c, "ToIgn3_5Unvalidated", options)
 	return cfg.(types.Config), r, err
 }
 
-// ToConfigBytes translates from a v4.22 Butane config to a v4.22 MachineConfig or a v3.6.0 Ignition config. It returns a report of any errors or
+// ToConfigBytes translates from a v4.22 Butane config to a v4.22 MachineConfig or a v3.5.0 Ignition config. It returns a report of any errors or
 // warnings in the source and resultant config. If the report has fatal errors or it encounters other problems
 // translating, an error is returned.
 func ToConfigBytes(input []byte, options common.TranslateBytesOptions) ([]byte, report.Report, error) {
 	if options.Raw {
-		return cutil.TranslateBytes(input, &Config{}, "ToIgn3_6", options)
+		return cutil.TranslateBytes(input, &Config{}, "ToIgn3_5", options)
 	} else {
 		return cutil.TranslateBytesYAML(input, &Config{}, "ToMachineConfig4_22", options)
 	}